pyproject.toml
[project]
name = "alma-item-checks-update-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Tom Boone",email = "tbc@tomboone.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "azure-functions (>=1.23.0,<2.0.0)",
    "wrlc-azure-storage-service (>=0.1.1,<0.2.0)",
    "wrlc-alma-api-client (>=0.1.7,<0.2.0)"
]

[[tool.poetry.source]]
name = "wrlc-python"
url = "https://pkgs.dev.azure.com/WRLCdev/Python/_packaging/wrlc-python/pypi/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
